changed location of dependency check to _lint.py
